BREAKING: Change `OffsetX`/`OffsetY` to float64 as used by TMX. Fixes

+++ tmx_group.go
@@ -37,9 +37,9 @@ type Group struct {
 	// The class of the group layer (since 1.9, defaults to "").
 	Class string `xml:"class,attr"`
 	// Rendering offset of the image layer in pixels. Defaults to 0. (since 0.15)
-	OffsetX int `xml:"offsetx,attr"`
+	OffsetX float64 `xml:"offsetx,attr"`
 	// Rendering offset of the image layer in pixels. Defaults to 0. (since 0.15)
-	OffsetY int `xml:"offsety,attr"`
+	OffsetY float64 `xml:"offsety,attr"`
 	// The opacity of the layer as a value from 0 to 1. Defaults to 1.
 	Opacity float32 `xml:"opacity,attr"`
 	// Whether the layer is shown (1) or hidden (0). Defaults to 1.
